Added foreign key from Ticket to User
Note that this work is not yet complete as I have commented out the
associations between Ticket and User to get around poison encoding
issues. I need to study up on this later. For now I will work with the
raw user_id column on the Elm side to continue working. This may actually
be the solution I want since this is an API-only application.